requirements.txt
flake8
pyyaml
charmhelpers
charms.reactive
